Move the models to the api package
