+++ api/api.go
@@ -3,7 +3,10 @@ package api
 import (
 	"database/sql"
 	_ "embed"
+	"encoding/json"
 	"fmt"
+	"io"
+	"net/http"
 	"os"
 	"path/filepath"
 
@@ -19,6 +22,42 @@ type Dictionary struct {
 	Word       string
 	Wordtype   string
 	Definition string
+}
+
+type Response struct {
+	Word       string     `json:"word"`
+	Phonetic   string     `json:"phonetic"`
+	Phonetics  Phonetics  `json:"phoenitcs"`
+	Meanings   []Meanings `json:"meanings"`
+	License    License    `json:"license"`
+	SourceUrls []string   `json:"sourceUrls"`
+}
+
+type Phonetics struct {
+	Text      string `json:"text"`
+	Audio     string `json:"audio"`
+	SourceURL string `json:"sourceUrl,omitempty"`
+	License   struct {
+		Name string `json:"name"`
+		URL  string `json:"url"`
+	} `json:"license,omitempty"`
+}
+
+type Meanings struct {
+	PartOfSpeech string `json:"partOfSpeech"`
+	Definitions  []struct {
+		Definition string `json:"definition"`
+		Synonyms   []any  `json:"synonyms"`
+		Antonyms   []any  `json:"antonyms"`
+		Example    string `json:"example,omitempty"`
+	} `json:"definitions"`
+	Synonyms []any `json:"synonyms"`
+	Antonyms []any `json:"antonyms"`
+}
+
+type License struct {
+	Name string `json:"name"`
+	URL  string `json:"url"`
 }
 
 func HandleLocalResponse(word string) ([]string, error) {
